// src/webpage/candidate-result-types.ts
// Generated by https://quicktype.io

export interface CandidateResults {
  kind:              string;
  url:               URL;
  queries:           Queries;
  context:           Context;
  searchInformation: SearchInformation;
  items:             Item[];
}

export interface Context {
  title: string;
}

export interface Item {
  kind:             Kind;
  title:            string;
  htmlTitle:        string;
  link:             string;
  displayLink:      DisplayLink;
  snippet:          string;
  htmlSnippet:      string;
  formattedUrl:     string;
  htmlFormattedUrl: string;
  pagemap:          Pagemap;
}

export enum DisplayLink {
  InLinkedinCOM = "in.linkedin.com",
  WWWLinkedinCOM = "www.linkedin.com",
}

export enum Kind {
  CustomsearchResult = "customsearch#result",
}

export interface Pagemap {
  cse_thumbnail: CSEThumbnail[];
  metatags:      Metatag[];
  cse_image:     CSEImage[];
  Person?:       Person[];
  ProfilePage?:  Person[];
}

export interface Person {
}

export interface CSEImage {
  src: string;
}

export interface CSEThumbnail {
  src:    string;
  width:  string;
  height: string;
}

export interface Metatag {
  "og:image":            string;
  "twitter:card":        TwitterCard;
  "linkedin:pagetag":    LinkedinPagetag;
  "platform-worker":     string;
  "al:android:package":  AlAndroidPackage;
  bingbot:               Bingbot;
  locale:                Locale;
  "al:ios:url":          string;
  "og:description":      string;
  "al:ios:app_store_id": string;
  platform:              string;
  "twitter:image":       string;
  "profile:last_name":   string;
  "twitter:site":        TwitterSite;
  litmsprofilename:      Litmsprofilename;
  "profile:first_name":  string;
  "og:type":             OgType;
  "twitter:title":       string;
  "al:ios:app_name":     AlAppName;
  "og:title":            string;
  pagekey:               Pagekey;
  "al:android:url":      string;
  viewport:              Viewport;
  "twitter:description": string;
  ubba:                  string;
  "og:url":              string;
  "al:android:app_name": AlAppName;
}

export enum AlAppName {
  LinkedIn = "LinkedIn",
}

export enum AlAndroidPackage {
  COMLinkedinAndroid = "com.linkedin.android",
}

export enum Bingbot {
  Nocache = "nocache",
}

export enum LinkedinPagetag {
  OpenToProvider = "openToProvider",
}

export enum Litmsprofilename {
  PublicProfileFrontend = "public-profile-frontend",
}

export enum Locale {
  EnUS = "en_US",
}

export enum OgType {
  Profile = "profile",
}

export enum Pagekey {
  PublicProfileV3Mobile = "public_profile_v3_mobile",
}

export enum TwitterCard {
  Summary = "summary",
}

export enum TwitterSite {
  Linkedin = "@Linkedin",
}

export enum Viewport {
  WidthDeviceWidthInitialScale10 = "width=device-width, initial-scale=1.0",
}

export interface Queries {
  request:  NextPage[];
  nextPage: NextPage[];
}

export interface NextPage {
  title:          string;
  totalResults:   string;
  searchTerms:    string;
  count:          number;
  startIndex:     number;
  inputEncoding:  string;
  outputEncoding: string;
  safe:           string;
  cx:             string;
}

export interface SearchInformation {
  searchTime:            number;
  formattedSearchTime:   string;
  totalResults:          string;
  formattedTotalResults: string;
}

export interface URL {
  type:     string;
  template: string;
}